requirements.txt
absl-py>=0.11.0
gin-config>=0.4.0
lingvo>=0.8.2
tensorflow-addons>=0.12.0
tensorflow-datasets>=4.2.0
